retruct repo to multiple folder

+++ install_pc.sh
@@ -1,10 +1,16 @@
 
+if [ -z ${ACTIVE_PROJECT} ]
+then
+    echo "Not found active project, please run init_env.sh <project>"
+    exit 1
+fi
+echo "START INSTALL PROJECT ${ACTIVE_PROJECT}"
 # remove stale nodes
 rm -f /dev/choen[0-1]
 # uninstall existing module
 rmmod choen
 # install module, exit if fail
-insmod $PWD/choen_dev/choen.ko || exit 1
+insmod $PWD/${ACTIVE_PROJECT}/choen.ko || exit 1
 # get major number from recently installed module
 major=`awk '/choen/ {print $1}' /proc/devices`
 # create new nodes
